requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
browser-use[cli]>=0.1.0
playwright>=1.48.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0
sqlmodel>=0.0.22
redis[hiredis]>=5.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic-settings>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3
langchain>=1.0.0
langgraph>=1.0.0
langchain-anthropic>=0.3.0
langchain-mcp-adapters>=0.1.0
